refactoring: Namespaces fixed for exceptions
