+++ model/import/Parser/Exception/InvalidCsvImportException.php
@@ -1,6 +1,7 @@
 <?php
 
 /*
+ *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; under version 2
@@ -15,12 +16,12 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2021 (original work) Open Assessment Technologies SA (under the project TAO-PRODUCT);
+ * Copyright (c) 2021  (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
 
-namespace oat\taoQtiItem\model\import\Parser;
+namespace oat\taoQtiItem\model\import\Parser\Exception;
 
 class InvalidCsvImportException extends InvalidImportException
 {
