+++ model/import/Parser/Exception/InvalidImportException.php
@@ -1,0 +1,63 @@
+<?php
+
+/*
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; under version 2
+ * of the License (non-upgradable).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Copyright (c) 2021  (original work) Open Assessment Technologies SA;
+ */
+
+declare(strict_types=1);
+
+namespace oat\taoQtiItem\model\import\Parser\Exception;
+
+use Exception;
+
+class InvalidImportException extends Exception
+{
+    private $totalErrors = 0;
+    private $messages = [];
+
+    protected const LEVEL = 1;
+
+    protected function addMessage(int $line, string $message, int $errorLevel): self
+    {
+        $this->message .= rtrim($message, ',').', ';
+
+        $this->messages[$errorLevel][] = [
+            'line' => $line,
+            'message' => $message,
+        ];
+
+        return $this;
+    }
+
+    public function addError(int $line, string $message): self
+    {
+        $this->totalErrors++;
+
+        return $this->addMessage($line, $message, static::LEVEL);
+    }
+
+    public function getErrors(): array
+    {
+        return $this->messages[static::LEVEL];
+    }
+
+    public function getTotalErrors(): int
+    {
+        return $this->totalErrors;
+    }
+}
