+++ model/import/Parser/Exception/WarningImportException.php
@@ -1,6 +1,7 @@
 <?php
 
 /*
+ *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; under version 2
@@ -15,48 +16,32 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2021 (original work) Open Assessment Technologies SA (under the project TAO-PRODUCT);
+ * Copyright (c) 2021  (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
 
-namespace oat\taoQtiItem\model\import\Parser;
+namespace oat\taoQtiItem\model\import\Parser\Exception;
 
-use Exception;
-
-class InvalidImportException extends Exception
+class WarningImportException extends InvalidImportException
 {
-    private $totalErrors = 0;
-    private $messages = [];
+    private $totalWarning = 0;
+    protected const LEVEL = 2;
 
-    protected const LEVEL = 1;
-
-    protected function addMessage(int $line, string $message, int $errorLevel): self
+    public function addWarning(int $line, string $message): self
     {
-        $this->message .= rtrim($message, ',').', ';
-
-        $this->messages[$errorLevel][] = [
-            'line' => $line,
-            'message' => $message,
-        ];
-
-        return $this;
-    }
-
-    public function addError(int $line, string $message): self
-    {
-        $this->totalErrors++;
+        $this->totalWarning++;
 
         return $this->addMessage($line, $message, static::LEVEL);
     }
 
-    public function getErrors(): array
+    public function getWarnings(): array
     {
-        return $this->messages[static::LEVEL];
+        return $this->message[static::LEVEL];
     }
 
-    public function getTotalErrors(): int
+    public function getTotalWarnings(): int
     {
-        return $this->totalErrors;
+        return $this->totalWarning;
     }
 }
