+++ model/import/Validator/Rule/QtiCompatibleXmlRule.php
@@ -21,7 +21,7 @@
 namespace oat\taoQtiItem\model\import\Validator\Rule;
 
 use oat\oatbox\service\ConfigurableService;
-use oat\taoQtiItem\model\import\Parser\RecoverableLineValidationException;
+use oat\taoQtiItem\model\import\Parser\Exception\RecoverableLineValidationException;
 
 class QtiCompatibleXmlRule extends ConfigurableService implements ValidationRuleInterface
 {
